Correção para deletar o usuário sem compras

+++ deleteUser.php
@@ -9,14 +9,27 @@ $daoUser = new DaoUser();
 
 $id_user = $_SESSION['id_user'];
 
-if ($daoSale->deleteSaleReference($id_user) > 0) {
-    if (($daoUser->deleteUser($id_user)) > 0) {
+$list = $daoSale->checkSale($id_user);
+$total_compras = $list[0];
+
+if ($total_compras['total_compra'] > 0) { // verifica se o usuario tem compras para remover as compras associadas
+    if ($daoSale->deleteSaleReference($id_user) > 0) { // remove todas as compras associadas ao usuario
+        if (($daoUser->deleteUser($id_user)) > 0) { // deleta o usuario
+            header('Location: index.php');
+            unset($_SESSION['id_user']);
+            exit();
+        } else {
+            echo "nao deletou no daoUser";
+        }
+    } else {
+        echo "nao deletou no daoSale";
+    }
+} else { // senão tiver compras, remove o usuario
+    if (($daoUser->deleteUser($id_user)) > 0) { 
         header('Location: index.php');
         unset($_SESSION['id_user']);
         exit();
     } else {
         echo "nao deletou no daoUser";
     }
-} else {
-    echo "nao deletou no daoSale";
 }
